Add tests for kwd/flag override
